requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
markupsafe==2.0.1

boto3~=1.28.36
Jinja2~=2.11.3
itsdangerous~=1.1.0
click~=7.1.2
PyJWT~=2.8.0
jmespath~=1.0.1
botocore~=1.31.36
s3transfer~=0.6.2
six~=1.16.0
python-dateutil~=2.8.2
urllib3~=1.26.16
